实现测试hmm

+++ note_op/src/note_op_util.cc
@@ -3,7 +3,7 @@
 int note_op_util::ArgMax(const double *data, int count) {
     // todo 这个能用cuda实现？
     double maxValue = -100000.0;
-    double maxIdx = 0;
+    int maxIdx = 0;
     for (int i = 0; i < count; ++i) {
         if (maxValue < data[i]) {
             maxValue = data[i];
